Clean up and sent functions to own files.
